% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{add_names}
\alias{add_names}
\title{Add in author and maintainer social media handles}
\usage{
add_names(df)
}
\arguments{
\item{df}{Data frame with \code{maintainer_name} and \code{author_github}}
}
\value{
Data frame with added names and social media handles.
}
\description{
Add in author and maintainer social media handles
}
\examples{
u <- uc_fetch() |>
  uc_fmt("2025-01-01") |>
  add_names()
}
