% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_users}
\alias{all_users}
\title{Fetch GH and Mastodon usernames}
\usage{
all_users(name, owner = "ropensci", pkg)
}
\arguments{
\item{name}{Character. Full name}

\item{owner}{Character. Repository owner.}

\item{pkg}{Character. Repository name (package name).}
}
\value{
Data frame of user names: gh_user and masto_user
}
\description{
Based on a name and a repository, fetch the usernames
}
\examples{
all_users(name = "Steffi LaZerte", pkg = "weathercan")
}
