% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{buff_bluesky}
\alias{buff_bluesky}
\title{ID for Buffer Bluesky channel}
\format{
An object of class \code{character} of length 1.
}
\usage{
buff_bluesky
}
\description{
ID for Buffer Bluesky channel
}
\keyword{datasets}
