% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{buff_channels}
\alias{buff_channels}
\title{List of IDs for Buffer channels: Bluesky, Mastodon, and LinkedIn}
\format{
An object of class \code{list} of length 3.
}
\usage{
buff_channels
}
\description{
List of IDs for Buffer channels: Bluesky, Mastodon, and LinkedIn
}
\keyword{datasets}
