% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{buff_linkedin}
\alias{buff_linkedin}
\title{ID for Buffer LinkedIn channel}
\format{
An object of class \code{character} of length 1.
}
\usage{
buff_linkedin
}
\description{
ID for Buffer LinkedIn channel
}
\keyword{datasets}
