% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{buff_mastodon}
\alias{buff_mastodon}
\title{ID for Buffer Mastodon channel}
\format{
An object of class \code{character} of length 1.
}
\usage{
buff_mastodon
}
\description{
ID for Buffer Mastodon channel
}
\keyword{datasets}
