% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{buff_nchars}
\alias{buff_nchars}
\title{Number of characters permitted by social media channels via Buffer}
\format{
An object of class \code{numeric} of length 3.
}
\usage{
buff_nchars
}
\description{
Number of characters permitted by social media channels via Buffer
}
\keyword{datasets}
