% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{buff_org}
\alias{buff_org}
\title{rOpenSci Buffer organization id}
\format{
An object of class \code{character} of length 1.
}
\usage{
buff_org
}
\description{
rOpenSci Buffer organization id
}
\keyword{datasets}
