% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_utils.R
\name{buffer_cleanup}
\alias{buffer_cleanup}
\title{Clean up draft/scheduled testing messages}
\usage{
buffer_cleanup()
}
\value{
Data frame with removed post ids
}
\description{
Removes all testing draft and/or scheduled messages matching "^testing(
scheduled)? Api again...$".
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
buffer_cleanup()
\dontshow{\}) # examplesIf}
}
