% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{buffer_org}
\alias{buffer_org}
\title{Return Buffer Organization account id}
\usage{
buffer_org(fields = "id", dry_run = FALSE)
}
\arguments{
\item{fields}{Character. Information to return. Defaults to 'id'.}

\item{dry_run}{Logical. Whether to do a dry run (i.e. don't talk to the API).}
}
\value{
The organization account id on Buffer.
}
\description{
Return Buffer Organization account id
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
buffer_org()
\dontshow{\}) # examplesIf}
}
