% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{buffer_posts_remove}
\alias{buffer_posts_remove}
\title{Remove posts on Buffer}
\usage{
buffer_posts_remove(id, dry_run = FALSE)
}
\arguments{
\item{id}{Character vector. Id(s) of the post(s) to be removed.}

\item{dry_run}{Logical. Whether to do a dry run (i.e. don't talk to the API).}
}
\value{
Data frame of removed posts
}
\description{
Remove posts on Buffer
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
p <- buffer_posts_write("testing Api again...", "now", channels = "mastodon")
buffer_posts_remove(p$id)
\dontshow{\}) # examplesIf}
}
