% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{by_platform}
\alias{by_platform}
\title{Arrange by platform}
\usage{
by_platform(df)
}
\arguments{
\item{df}{Data frame. Formatted data including social media handles.}
}
\value{
Data frame arranged by platform on which to advertise.
}
\description{
Arranges data in long by platform (linkedin and mastodon).
}
\examples{
u <- uc_fetch() |>
  uc_fmt("2025-01-01") |>
  uc_handles() |>
  by_platform()
}
