% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_dir}
\alias{cache_dir}
\title{Cache folder to store data}
\usage{
cache_dir(type)
}
\arguments{
\item{type}{Character. Data type to cache.}
}
\value{
Character file path
}
\description{
Where the cache will be or is for specific types of data cached by
promoutils.
}
\examples{
cache_dir(type = "matomo")
cache_dir(type = "help-wanted")
}
