% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chats.R
\name{chats_announce}
\alias{chats_announce}
\title{Announce/remind people about chats}
\usage{
chats_announce(
  when = NULL,
  channel = "coffee-chats",
  dry_run = TRUE,
  test_run = FALSE
)
}
\arguments{
\item{when}{Character or Date/time. When to post announcement.}

\item{channel}{Character. Channel name to make announcement in.}

\item{dry_run}{Logical. Whether to do a dry run (i.e. don't talk to the API).}

\item{test_run}{Logical. Whether to do a test run (i.e. post to a test area).}
}
\value{
Success message
}
\description{
Announce/remind people about chats
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
chats_announce(when = "now", test_run = TRUE)
\dontshow{\}) # examplesIf}
}
\seealso{
Slacks Special Mentions - \url{https://docs.slack.dev/messaging/formatting-message-text/#special-mentions}
}
