% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chats.R
\name{chats_signups}
\alias{chats_signups}
\title{Get chat sign ups}
\usage{
chats_signups(channel = "coffee-chats", author_lang = "English")
}
\arguments{
\item{channel}{Character. Channel Name. Defaults to "coffee-chats"}

\item{author_lang}{Character vector. If the author of the announcement post
wishes to participate, these are the languages to include for them. Any of
\code{chats_languages()}.}
}
\value{
Data frame of users and languages, with one row per user per
language.
}
\description{
Returns the users who reacted to the most recent chat announcement (see
\code{chats_announce()}) with a language emoji.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
chats_signups(channel = "testing-api")
\dontshow{\}) # examplesIf}
}
\references{
https://docs.slack.dev/reference/methods/reactions.get
}
