% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coworking.R
\name{cw_checkin}
\alias{cw_checkin}
\title{Create draft message for checking in with Cohosts}
\usage{
cw_checkin(which = "next", names = NULL, notes_link, slides_link)
}
\arguments{
\item{which}{Character/Date. "next" to fetch details on the next coworking
session, "last" to fetch details on the last scheduled (in future)
coworking session, or a Date fetch details for a specific coworking
session.}

\item{names}{Character. Names of cohost if overriding those in the event listing.}

\item{notes_link}{Character. Link to the Google doc with coworking notes.}

\item{slides_link}{Character. Link to the Google slides.}
}
\description{
The week before, prepare the slides and coworking document, then use this
draft text to invite the cohost(s) to review via Slack or Email.
}
