% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coworking.R
\name{cw_checkin_event}
\alias{cw_checkin_event}
\title{Create draft message for checking in with Cohosts about the Event review}
\usage{
cw_checkin_event(which = "next", names = NULL, print = FALSE)
}
\arguments{
\item{which}{Character/Date. "next" to fetch details on the next coworking
session, "last" to fetch details on the last scheduled (in future)
coworking session, or a Date fetch details for a specific coworking
session.}

\item{names}{Character. Names of cohost if overriding those in the event listing.}

\item{print}{Logical. Whether to print the copied text to the console.}
}
\description{
When an event has been drafted use this
draft text to invite the cohost(s) to review via Slack or Email.
}
\examples{
# cw_checkin_event("2025-12-02")
cw_checkin_event("2026-04-07")
}
