% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coworking.R
\name{cw_details}
\alias{cw_details}
\title{Fetch details about coworking sessions}
\usage{
cw_details(which = "next")
}
\arguments{
\item{which}{Character/Date. "next" to fetch details on the next coworking
session or a Date fetch details for a specific coworking session.}
}
\value{
Data frame with coworking event details
}
\description{
Fetch details about coworking sessions
}
\examples{
cw_details()
cw_details("2023-11")
}
