% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coworking.R
\name{cw_docs_link}
\alias{cw_docs_link}
\title{Get link to Coworking docs}
\usage{
cw_docs_link(which = "next", open_sites = TRUE, add = FALSE)
}
\arguments{
\item{which}{Character/Date. "next" to fetch details on the next coworking
session, "last" to fetch details on the last scheduled (in future)
coworking session, or a Date fetch details for a specific coworking
session.}

\item{open_sites}{Logical. Open websites with relevant details?}

\item{add}{Logical. Whether to initialize a PR and add the link to the events
file.}
}
\value{
Google Docs link
}
\description{
Creates the coworking doc if it doesn't exist and returns a link either way.
Optionally creates PR adds link to event page.
}
