% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coworking.R
\name{cw_event}
\alias{cw_event}
\title{Create a draft event for coworking}
\usage{
cw_event(date)
}
\arguments{
\item{date}{Character/Date. Date of the coworking event (local)}

\item{tz}{Character. Timezone of the event (match from \code{OlsonNames()})}

\item{theme}{Character. Theme.}

\item{cohost}{Character. Name of cohost}
}
\description{
Creates a draft coworking event for the roweb3 website
}
