% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coworking.R
\name{cw_issue}
\alias{cw_issue}
\title{Create Coworking To-Dos}
\usage{
cw_issue(
  date = NULL,
  tz = NULL,
  theme = "XXXX",
  cohost = "XXXX",
  repo = "rosadmin/comms",
  dry_run = FALSE
)
}
\arguments{
\item{date}{Character. Date of next event (if NULL picks next first Tuesday).}

\item{tz}{Character. Timezone of next event (if NULL picks next in order).}

\item{theme}{Character. Name of theme, if unknown, uses XXXX placeholder}

\item{cohost}{Character. Name of cohost, if unknown, uses XXXX placeholder}

\item{repo}{Character. GitHub repository including owner (e.g.,
\code{rosadmin/comms})}

\item{dry_run}{Logical. Whether to really create the issue or just return the
text.}
}
\value{
Nothing
}
\description{
Create a GitHub issue in a repository listing the Coworking ToDos. If no date
or timezone, picks the next appropriate date (first Tuesday of the month
following an existing coworking issue) and next appropriate timezone (cycling
through America, Europe, and Australia) automatically.
}
\examples{
cw_issue(dry_run = TRUE)
}
