% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coworking.R
\name{cw_posts}
\alias{cw_posts}
\title{Create a draft post for coworking}
\usage{
cw_posts(
  date,
  host = list(mastodon = "@steffilazerte@fosstodon.org", slack = "UNRAUCMTK", linkedin =
    "Steffi LaZerte", bluesky = "@steffilazerte.bsky.social"),
  cohost,
  posters_tz = Sys.timezone(),
  test_run = FALSE,
  dry_run = FALSE,
  print = TRUE,
  branch = NULL
)
}
\arguments{
\item{date}{Character/Date. Date of the coworking event (local)}

\item{host}{List. Named list of social media handles for the host, must include
"mastodon", "bluesky", "linkedin", and "slack". Where there is no handle,
use the host's full name.}

\item{cohost}{List. Named list of social media handles for the cohost. must include
"mastodon", "bluesky", "linkedin", and "slack". Where there is no handle,
use the cohost's full name.}

\item{posters_tz}{Character. Timezone of poster. Required for getting the
time at which to post Slack messages as these are posted in the local
timezone.}

\item{test_run}{Logical. Whether to do a test run (i.e. post to a test area).}

\item{dry_run}{Logical. Whether to do a dry run (i.e. don't talk to the API).}

\item{print}{Logical. Whether to simply print the text to console instead of
copying to the clipboard.}

\item{branch}{Character. Branch name if not on main.}
}
\description{
Creates draft posts for Bluesky, Mastodon, and LinkedIn (through Buffer) and
Slack (through Slack API, and text for copy/paste to other Slack spaces).
}
\examples{
cw_posts(
  "2023-07-04",
  cohost = list(
    "mastodon" = "@cohost@mastodon.org",
    "bluesky" = "@cohost@bsky.social",
    "linkedin" = "Cohost",
    "slack" = "UXXXXXXX"
  ),
  dry_run = TRUE
)

\dontrun{
  cw_posts(
    "2023-07-04",
     cohost = list(
      "mastodon" = "@cohost@mastodon.org",
      "bluesky" = "@cohost@bsky.social",
      "linkedin" = "Cohost",
      slack = "UXXXXXXX"
    )
  )
}
}
