% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coworking.R
\name{cw_slides_link}
\alias{cw_slides_link}
\title{Open and fetch link to coworking slides}
\usage{
cw_slides_link(open_site = TRUE)
}
\arguments{
\item{open_site}{Logical. Open slides in browser?}
}
\value{
Link to slides
}
\description{
Open and fetch link to coworking slides
}
