% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coworking.R
\name{cw_todos_list}
\alias{cw_todos_list}
\title{Create a text based todo list for Coworking tasks}
\usage{
cw_todos_list(start = NULL, end = NULL)
}
\arguments{
\item{start}{Start date for coworking tasks. \code{NULL} by default to do Jan-Dec
for the next year.}

\item{end}{End date for coworking tasks. \code{NULL} by default to do Jan-Dec for
the next year.}
}
\value{
Text TO-DO list
}
\description{
This list can be copy/pasted into todo software to keep track of dated
reminders for arranging coworking events.
}
\examples{
cw_todos_list()
}
