% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coworking.R
\name{cw_tz}
\alias{cw_tz}
\title{Work with Coworking timezones}
\usage{
cw_tz(tz = NULL)
}
\arguments{
\item{tz}{Character. Timezone to switch the display type of. If NULL, return
all timezones.}
}
\value{
Character of the opposite timezone display type, or a vector of all
timezones.
}
\description{
Transform or return coworking timezones.
}
\examples{
cw_tz()
cw_tz("Americas Pacific")
cw_tz("America/Vancouver")
cw_tz("Europe Central")
}
