% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{dict_helpwanted}
\alias{dict_helpwanted}
\title{Dictionary for help-wanted}
\usage{
dict_helpwanted()
}
\value{
Data frame.
}
\description{
Creates a data frame dictionary of different versions of wording to use in
help-wanted posts depending on the language of the post.
}
\examples{
dict_helpwanted()
}
