% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{discourse_user}
\alias{discourse_user}
\title{Fetch full name of Discourse user (id)}
\usage{
discourse_user(user)
}
\arguments{
\item{user_id}{Integer. Discourse user id.}
}
\value{
Character
}
\description{
Using the Discourse username id, return the full name of that discourse user.
}
\examples{
# discourse_user(1) # Requires authentication
}
