% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_templates.R
\name{event_coworking}
\alias{event_coworking}
\title{Create a template event for coworking}
\usage{
event_coworking(date, tz, theme = "THEME", cohost = "COHOST")
}
\arguments{
\item{date}{Character/Date. Date of the coworking event (local)}

\item{tz}{Character. Timezone of the event (match from \code{OlsonNames()})}

\item{theme}{Character. Theme.}

\item{cohost}{Character. Name of cohost}
}
\description{
Create a template event for coworking
}
