% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{forum_mention}
\alias{forum_mention}
\title{Extract mentions from forum text}
\usage{
forum_mention(x)
}
\arguments{
\item{x}{Forum text}
}
\value{
Character of metions
}
\description{
Extract mentions from forum text
}
