% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{forum_post}
\alias{forum_post}
\title{Fetch post text from by topic id}
\usage{
forum_post(topic_id)
}
\arguments{
\item{x}{Topic id}
}
\value{
HTML of the post
}
\description{
Fetch post text from by topic id
}
\examples{
# forum_post(3920) # Needs auth
}
