% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{forum_resource}
\alias{forum_resource}
\title{Extract resources from forum text}
\usage{
forum_resource(x)
}
\arguments{
\item{x}{Forum text}
}
\value{
Character vector of resources
}
\description{
Extract resources from forum text
}
