% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gh_cache}
\alias{gh_cache}
\title{Create a cached version of the GH api calls}
\usage{
gh_cache(
  endpoint,
  ...,
  per_page = NULL,
  .per_page = NULL,
  .token = NULL,
  .destfile = NULL,
  .overwrite = FALSE,
  .api_url = NULL,
  .method = "GET",
  .limit = NULL,
  .accept = "application/vnd.github.v3+json",
  .send_headers = NULL,
  .progress = TRUE,
  .params = list(),
  .max_wait = 600,
  .max_rate = NULL
)
}
\description{
Create a cached version of the GH api calls
}
\details{
\code{memoise::memoise(gh::gh)}
}
