% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_issues.R
\name{gh_issue_close}
\alias{gh_issue_close}
\title{Close a GH issue}
\usage{
gh_issue_close(n, owner, repo, dry_run = FALSE)
}
\arguments{
\item{n}{Numeric. Issue number to close.}

\item{owner}{Character. GitHub username or organization owner of the repository.}

\item{repo}{Character. Repository in which to create issue.}

\item{dry_run}{Logical. Whether to do a dry run (i.e. don't talk to the API).}
}
\description{
Closes a GH issue by number, repository and owner.
}
\examples{
\dontrun{
  # Closes the first issue in the "myrop" repository of "myhandle" on GitHub
  gh_issue_close(1, "myhandle", "myrepo")
}
}
