% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_issues.R
\name{gh_issue_fetch}
\alias{gh_issue_fetch}
\title{Fetch issues from a GH repository}
\usage{
gh_issue_fetch(
  state = "open",
  labels = NULL,
  since = NULL,
  owner = "rosadmin",
  repo = "scheduled_socials",
  issue = NULL,
  verbose = FALSE
)
}
\arguments{
\item{state}{Character. Which issues to fetch: "open", "closed", "all"}

\item{labels}{Character vector. Fetch only issues with these labels. (Optional)}

\item{since}{Character/Date/datetime. Fetch only issues since this date/time. (Optional)}

\item{owner}{Character. Owner of the repository}

\item{repo}{Character. Name of the repository (name of the package)}

\item{issue}{Numeric. Specific Issue number to fetch.}

\item{verbose}{Logical. Show progress messages.}
}
\value{
List of issues
}
\description{
Fetch issues from a GH repository
}
\examples{
i <- gh_issue_fetch()
}
