% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{gh_issue_fmt}
\alias{gh_issue_fmt}
\title{Format issues list from GH}
\usage{
gh_issue_fmt(
  i,
  which = c("title", "number", "body", "labels", "url", "created", "updated",
    "gh_user_issue")
)
}
\arguments{
\item{i}{List of issues from \code{gh_issue_fetch()}}

\item{which}{Which fields to includ}
}
\value{
Issues formated as a data frame
}
\description{
Format issues list from GH
}
\examples{
i <- gh_issue_fetch()
i <- gh_issue_fmt(i, which = "title")
}
