% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{gh_issue_labels}
\alias{gh_issue_labels}
\title{Extract information on help-wanted labels for issues}
\usage{
gh_issue_labels(
  i,
  labels_help = "(help)|(help wanted)|(help-wanted)|(help_wanted)",
  labels_first = "(good first issue)|(beginner)|(good-first-issue)"
)
}
\arguments{
\item{i}{Data frame of issues from \code{gh_issue_fmt()}}

\item{labels_first}{Character. Regular expression to meatch good-first-issue
labels.}

\item{labels_help.}{Character. Regular expression to match help-wanted
labels.}
}
\value{
data frame with added label details
}
\description{
Extract information on help-wanted labels for issues
}
\examples{
i <- gh_issue_fetch(owner = "ropensci", repo = "weathercan")
i <- gh_issue_fmt(i)
i <- gh_issue_labels(i)
}
