% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_issues.R
\name{gh_issue_post}
\alias{gh_issue_post}
\title{Create a GH issues post}
\usage{
gh_issue_post(
  title,
  body,
  labels,
  owner,
  repo,
  avoid_dups = TRUE,
  dry_run = FALSE,
  open_browser = TRUE
)
}
\arguments{
\item{title}{Character. Title of the issue.}

\item{body}{Character. Body text of the issue.}

\item{labels}{Character vector. Vector of labels to assign.}

\item{owner}{Character. GitHub username or organization owner of the repository.}

\item{repo}{Character. Repository in which to create issue.}

\item{avoid_dups}{Logical. If TRUE (default), avoids posting duplicate issues.}

\item{dry_run}{Logical. Whether to do a dry run (i.e. don't talk to the API).}

\item{open_browser}{Logical. Whether to open the issue in the browser.}
}
\description{
Opens an issue on Github with title, body and labels.
}
\examples{
gh_issue_post(
  title = "My issue",
  body = "Body of the issue",
  labels = "help-wanted",
  owner = "ropensci",
  repo = "weathercan",
  dry_run = TRUE
)
}
