% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_issues.R
\name{gh_label_events}
\alias{gh_label_events}
\title{Fetch details of specific labels for issues}
\usage{
gh_label_events(owner, repo, issue, labels)
}
\arguments{
\item{owner}{Character. Owner of the repository}

\item{repo}{Character. Name of the repository (name of the package)}

\item{issue}{Numeric. Specific Issue number to fetch.}

\item{labels}{Character vector. Fetch only these labels.}
}
\value{
tibble with gh user name for the labeller and the date the label was
created.
}
\description{
Fetch details of specific labels for issues
}
\examples{
gh_label_events("ropensci", "weathercan", issue = 149, labels = "help wanted")
}
