% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gh_masto}
\alias{gh_masto}
\title{Find Mastodon username from GitHub profile}
\usage{
gh_masto(gh_user)
}
\arguments{
\item{gh_user}{Character. GH username}
}
\value{
Character URL to mastodon profile if it exists, NA otherwise.
}
\description{
Find Mastodon username from GitHub profile
}
\examples{
gh_masto("steffilazerte")
}
