% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gh_name}
\alias{gh_name}
\title{Fetch full name of GitHub user}
\usage{
gh_name(gh_user)
}
\arguments{
\item{gh_user}{Character. GitHub username/handle}
}
\value{
Character, full name of the user or NA
}
\description{
Fetch full name of GitHub user
}
\examples{
gh_name("steffilazerte")
}
