% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{gh_user}
\alias{gh_user}
\title{Find GH username from repository and full name}
\usage{
gh_user(name, owner = "ropensci", pkg, .max_rate = NULL)
}
\arguments{
\item{name}{Character. Full or partial name of the person for whom you want
to fetch the GitHub username.}

\item{owner}{Character. Owner of the repository.}

\item{pkg}{Character. Repository name (package name).}

\item{.max_rate}{Numeric. Passed through to \code{gh:gh()}.}
}
\value{
Data frame with names attempted and usernames found
}
\description{
Look up users of a repository and match to a name. Try with and without
initials.
}
\examples{

gh_user(name = "Steffi E. LaZerte", owner = "ropensci", pkg = "weathercan")
gh_user(name = "Steffi", owner = "ropensci", pkg = "weathercan")
}
