% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_wanted.R
\name{help_handles}
\alias{help_handles}
\title{Get a data frame of social media handles for help-wanted issues}
\usage{
help_handles(help)
}
\arguments{
\item{help}{Data frame of help wanted issues.}
}
\value{
Data frame of social media handles associated with \code{help}
}
\description{
Get a data frame of social media handles for help-wanted issues
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
help_fetch() |>
  help_handles()
\dontshow{\}) # examplesIf}
}
