% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_wanted.R
\name{help_preview}
\alias{help_preview}
\title{Prepare help-wanted posts}
\usage{
help_preview(help)
}
\arguments{
\item{help}{Data frame. Formatted help-wanted issues including social media
handles.}
}
\value{
Copies commands to clipboard, optionally prints if (\code{print = TRUE}).
}
\description{
Prepare posts and preview
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
h <- help_fetch()
help_preview(h)
\dontshow{\}) # examplesIf}
}
