% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_wanted_json.R
\name{help_wanted_json}
\alias{help_wanted_json}
\title{Create help wanted JSON file}
\usage{
help_wanted_json()
}
\value{
Creates/updates \code{issues.json}
}
\description{
Function created for rosadmin/help-wanted. Fetches ropensci repos with
issues labelled help-wanted. Saves output to issues.json for use by
helpwanted workflows including https://ropensci.org/help-wanted.
}
