% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_wanted_json.R
\name{hw_issues}
\alias{hw_issues}
\title{Create help wanted JSON file}
\usage{
hw_issues(
  min_date = Sys.Date() - lubridate::years(2),
  throttle = 10/60,
  path = "issues.json",
  verbose = TRUE
)
}
\arguments{
\item{min_date}{Date. Fetch issues since this data.}

\item{throttle}{Numeric. Maximum requests per second. Default throttled to 10
requests per minute which is the max rate for unauthenticated GitHub Pats
(i.e. on GH actions).}

\item{path}{Character. Filename/path to save json of issues to.}

\item{verbose}{Logical. Show extra informative messages.}
}
\value{
Creates/updates issues json list at \code{path}
}
\description{
Function created for rosadmin/help-wanted. Fetches package repos with
issues labelled help-wanted (and variants). Saves output to issues.json for
use by helpwanted workflows including https://ropensci.org/help-wanted.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
hw_issues()
\dontshow{\}) # examplesIf}
}
