% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{keys_check}
\alias{keys_check}
\title{Key status of credentials}
\usage{
keys_check()
}
\value{
Message informing of status
}
\description{
Check to see if you have set up all the keys required by promoutils.
}
\examples{
keys_check()
}
