% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{keys_set}
\alias{keys_set}
\title{Set keys required for promoutils API access}
\usage{
keys_set(type = NULL)
}
\arguments{
\item{type}{Character Vector. Keys/tokens to set ("slack", "matomo",
"linkedin", "github").}
}
\value{
Nothing, side effect of setting token credentials in the keyring or
via \code{gitcreds::gitcres_set()} for GitHub tokens.
}
\description{
This function guides uses through the finding and setting tokens so that
promoutils can find them.
}
