% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{li_org}
\alias{li_org}
\title{rOpenSci linkedin organization URN}
\format{
An object of class \code{character} of length 1.
}
\usage{
li_org
}
\description{
rOpenSci linkedin organization URN
}
\keyword{datasets}
