% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedin.R
\name{li_posts_write}
\alias{li_posts_write}
\title{Post to LinkedIn}
\usage{
li_posts_write(author, body, dry_run = FALSE)
}
\arguments{
\item{author}{Character. URN. Either yours (see \code{li_urn_me()} or rOpenSci's
"urn:li:organization:77132573")}

\item{body}{Character. The body of the post as you would like it to appear.}

\item{dry_run}{Logical. TRUE to show what would be sent to the server without
actually sending it.}
}
\value{
A string of the URN for the post id.
}
\description{
Post to LinkedIn
}
\examples{
# Dry-run
id <- li_posts_write(
  author = li_org, # Post on behalf of rOpenSci
  body = "Testing out the LinkedIn API via R and httr2!",
  dry_run = TRUE)

\dontrun{
# Real post
id <- li_posts_write(
  author = li_org, # Post on behalf of rOpenSci
  body = "Testing out the LinkedIn API via R and httr2!")
}
}
