% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedin.R
\name{li_urn_me}
\alias{li_urn_me}
\title{Fetch your personal URN number}
\usage{
li_urn_me()
}
\value{
A string with your URN in the format of "urn:li:person:XXXX"
}
\description{
This is required to post on LinkedIn to your personal account
(for rOpenSci, use the organization urn, "urn:li:organization:77132573"
}
\examples{

\dontrun{
li_urn_me()
}
}
