% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matomo.R
\name{matomo_all}
\alias{matomo_all}
\title{Download all Matomo historic views}
\usage{
matomo_all(year_range = 2010:2024, host = "https://ropensci.matomo.cloud")
}
\arguments{
\item{year_range}{Numeric vector. Years to download views for.}

\item{host}{Character. URL for the Matomo host to collect views for. Defaults
to that of rOpenSci.}
}
\value{
Nothing. Writes views to disk.
}
\description{
Downloads all views from a specific year range (2010-2024 by defautl).
Views can be read with \code{\link[=matomo_read]{matomo_read()}}.
}
\details{
Requires a Renviron "MATOMO_TOKEN" key representing your token for Matomo
for the host you'd like to fetch data for.
}
