% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matomo.R
\name{matomo_blogposts}
\alias{matomo_blogposts}
\title{Formats and filter Matomo views to blogposts}
\usage{
matomo_blogposts(views)
}
\arguments{
\item{views}{Data frame from \code{\link[=matomo_read]{matomo_read()}}.}
}
\value{
Data frame of blog post views
}
\description{
Formats and filter Matomo views to blogposts
}
\examples{
\dontshow{if (dir.exists(cache_dir("matomo"))) withAutoprint(\{ # examplesIf}
matomo_read() |>
  matomo_blogposts()
\dontshow{\}) # examplesIf}
}
