% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matomo.R
\name{matomo_read}
\alias{matomo_read}
\title{Read Matomo views saved to disk}
\usage{
matomo_read()
}
\value{
Data frame of view data.
}
\description{
Reads a saved set of views.
}
\examples{
\dontshow{if (dir.exists(matomo_dir())) withAutoprint(\{ # examplesIf}
matomo_read()
\dontshow{\}) # examplesIf}
}
