% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matomo.R
\name{matomo_update}
\alias{matomo_update}
\title{Update Matomo views}
\usage{
matomo_update(
  year = lubridate::year(Sys.Date()),
  host = "https://ropensci.matomo.cloud"
)
}
\arguments{
\item{year}{Numeric. Year to update}

\item{host}{Character. URL for the Matomo host to collect views for. Defaults
to that of rOpenSci.}
}
\value{
Nothing. Updates views to disk.
}
\description{
Update a specific year of Matomo blog views. Defaults to current year.
Views can be read with \code{\link[=matomo_read]{matomo_read()}}.
}
\details{
Requires a .Renviron "MATOMO_TOKEN" key representing your token for Matomo
for the host you'd like to fetch data for.
}
