% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkgs}
\alias{pkgs}
\title{Return a data frame of rOpenSci packages}
\usage{
pkgs(
  url = "https://ropensci.github.io/roregistry/registry.json",
  which = "active",
  return = "sub"
)
}
\arguments{
\item{url}{Character. Registry url}

\item{which}{Character. Status of packages to return ("all" or "active")}

\item{return}{Character. Return a subset ("sub") or all ("all") package fields.}
}
\value{
data frame
}
\description{
Return a data frame of rOpenSci packages
}
\examples{
pkgs()
pkgs(which = "all", return = "all")
}
