% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkgs_ru}
\alias{pkgs_ru}
\title{List of packages and details from R-Universe API}
\usage{
pkgs_ru(universe = "ropensci")
}
\arguments{
\item{universe}{Character. Universe to collect details from.}
}
\description{
List of packages and details from R-Universe API
}
\examples{
pkgs_ru()
}
