% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_templates.R
\name{post_coworking}
\alias{post_coworking}
\title{Create a template post for coworking}
\usage{
post_coworking(date, when = "week", who)
}
\arguments{
\item{date}{Character/Date. Date of the coworking event (local)}

\item{when}{Character. Is this a "week" before reminder or an "hour" before reminder?}

\item{who}{Character. The full mastodon handle for the cohost (i.e. XXXX@XXXX.com)}
}
\description{
Create a template post for coworking
}
