% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socials.R
\name{post_issue}
\alias{post_issue}
\title{Create a draft issue to post to Mastodon and LinkedIn}
\usage{
post_issue(
  time,
  tz,
  title,
  body,
  where = "mastodon",
  dry_run = FALSE,
  avoid_dups = TRUE
)
}
\arguments{
\item{time}{Date/time. Date and time at which the post should be made}

\item{tz}{Character. Timezone (from \code{OlsonNames()}) in which to post}

\item{title}{Character. Title of the post ("\link{Post}" and the date will be
prepended and appended}

\item{body}{Character. Text to be posted (omit the YAML for posting info;
#RStats and @rstats@a.gup.pe will be appended}

\item{where}{Character vector. Either \code{mastodon} and/or \code{linkedin} to
specifty which platforms this should be posted on.}

\item{dry_run}{Logical. Whether to perform a dry run (do not post, but
display draft).}

\item{avoid_dups}{Logical. Don't post an issue if any open issue has the
same title.}
}
\description{
Formats the body and title of an issue and posts it on
"rosadmin/scheduled_socials". The issue will bed opened in a browser for
editing and confirmation. Note that issues will not be posted until the
labels "draft" and "needs-review" have been removed.
}
