% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{post_time}
\alias{post_time}
\title{Get the next post date/time}
\usage{
post_time(day, hour)
}
\arguments{
\item{day}{Character. Day of the week (e.g., "Monday")}

\item{hour}{Numeric. Hour at which to post (e.g., 8)}
}
\value{
Date time as character (without timezone)
}
\description{
Finds the next date/time to post by day of the week and hour.
}
\examples{
post_time("Wednesday", 8)
}
