% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promoutils-package.R
\docType{package}
\name{promoutils-package}
\alias{promoutils}
\alias{promoutils-package}
\title{promoutils: Utilities for Promoting rOpenSci}
\description{
Utility functions and workflows for promoting community events and community members on rOpenSci website and social media.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/promoutils}
  \item \url{https://github.com/ropensci-org/promoutils}
  \item Report bugs at \url{https://github.com/ropensci-org/promoutils/issues}
}

}
\author{
\strong{Maintainer}: Steffi LaZerte \email{sel@steffilazerte.ca} (\href{https://orcid.org/0000-0002-7690-8360}{ORCID})

Other contributors:
\itemize{
  \item Yanina Bellini Saibene \email{yabellini@ropensci.org} (\href{https://orcid.org/0000-0002-4522-7466}{ORCID}) [contributor]
}

}
\keyword{internal}
