% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prs_list}
\alias{prs_list}
\title{List PRs}
\usage{
prs_list(match = NULL, owner = "ropensci", repo = "roweb3")
}
\arguments{
\item{match}{Character. String to match in the title or branch name}

\item{owner}{Character. GitHub owner of the repository (defaults to 'ropensci')}

\item{repo}{Character. GitHub repository name (defaults to 'roweb3')}
}
\value{
Data frame with PR url, number, title, and branch name ('ref')
}
\description{
List open PRs by url, title and number, optionally matching to a title or branch name (ref)
}
\examples{
prs_list("coworking") # List all coworking related (open) PRs
}
