% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ro_masto}
\alias{ro_masto}
\title{Find Mastodon username from rOpenSci author pages}
\usage{
ro_masto(name)
}
\arguments{
\item{gh_user}{Character. Full name as on RO author pages}
}
\value{
Character URL to mastodon profile if it exists, NA otherwise.
}
\description{
Find Mastodon username from rOpenSci author pages
}
\examples{
ro_masto("Steffi LaZerte")
}
