% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ro_urn}
\alias{ro_urn}
\title{rOpenSci linkedin organization URN}
\format{
An object of class \code{character} of length 1.
}
\usage{
ro_urn
}
\description{
rOpenSci linkedin organization URN
}
\keyword{datasets}
