% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{slack_cleanup}
\alias{slack_cleanup}
\title{Clean up old scheduled messages}
\usage{
slack_cleanup()
}
\description{
Removes previously scheduled messages from \verb{#admin-scheduled} if after the
posting date.
}
\examples{

slack_posts_write("testing cleanup",
                  when = Sys.time() + lubridate::seconds(600),
                  tz = Sys.timezone())
slack_scheduled_list()

slack_cleanup()
}
