% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{slack_message_rm}
\alias{slack_message_rm}
\title{Remove a Slack message}
\usage{
slack_message_rm(msg = NULL, channel = NULL, channel_id = NULL, ts)
}
\arguments{
\item{msg}{Data frame. Output of \code{slack_messages} containing messages to
remove. Should contain columns "channel" and "ts". Note that only the most
recent message will be removed.}

\item{channel}{Character. Channel Name. Not required if channel_id supplied.}

\item{channel_id}{Character. Channel id.}

\item{ts}{Numeric. Timestamp to identify message to remove}
}
\value{
Nothing. Side effect of removing message from channel
}
\description{
Use with caution!
}
