% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{slack_message_rm_bulk}
\alias{slack_message_rm_bulk}
\title{Bulk remove Slack messages}
\usage{
slack_message_rm_bulk(msg, channel = NULL, channel_id = NULL, ts = NULL)
}
\arguments{
\item{msg}{Data frame. Output of \code{slack_messages} containing messages to
remove. Should contain columns "channel" and "ts". Removes all messages.}

\item{channel}{Character. Channel Name. Not required if channel_id supplied.}

\item{channel_id}{Character. Channel id.}

\item{ts}{Numeric. Timestamp to identify message to remove}
}
\value{
Nothing. Side effect of removing message from channel
}
\description{
Use with caution! This removes all messages matched by timestamp \code{ts} in
\code{channel} or \code{channel_id}. Note that users can only delete their own messages.
}
