% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{slack_permalink}
\alias{slack_permalink}
\title{Create a permalink for a Slack message}
\usage{
slack_permalink(ts, channel = NULL)
}
\arguments{
\item{ts}{Numeric. Timestamp to identify the messages}

\item{channel}{Character vector. Channel(s) to post message to.}
}
\value{
Character URL to Slack message
}
\description{
Create a permalink for a Slack message
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
m <- slack_messages("coffee-chats")
slack_permalink(ts = m$ts[1], channel = m$channel[1])
\dontshow{\}) # examplesIf}
}
