% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{slack_posts_write}
\alias{slack_posts_write}
\title{Write Slack message}
\usage{
slack_posts_write(
  body,
  when = "now",
  tz = "America/Winnipeg",
  channel = "#testing-api",
  dry_run = FALSE
)
}
\arguments{
\item{body}{Character. Text of message to post.}

\item{when}{Character or Date/time. When to post message (timezone is
ignored, see \code{tz}).}

\item{tz}{Character. Timezone of \code{when}.}

\item{channel}{Character. Channel to post message to.}

\item{dry_run}{Logical. Test run?}
}
\value{
Success message
}
\description{
Write a Slack message for posting now or later.
}
\details{
See https://docs.slack.dev/messaging/formatting-message-text#special-mentions
https://docs.slack.dev/messaging/formatting-message-text#mentioning-users
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
slack_posts_write("testing on Tuesday")
slack_posts_write("testing more and more", when = Sys.time() + 3600, tz = "Europe/Paris")
slack_posts_write(
  paste(
  "Join us for Social Coworking and office hours next week!",
  "",
  ":grey_exclamation: Theme: TESTING",
  ":hourglass_flowing_sand: When: TODAY!",
  ":cookie: Hosted by: USER! and cohost HOST",
  "",
  "You can use this time for...",
  "- General coworking", sep = "\n"), when = "now")

# Dry runs
slack_posts_write("testing on Tuesday", dry_run = TRUE)
slack_posts_write(
  "testing [this cool link](https://mycoolsite.com)",
  dry_run = TRUE
)
\dontshow{\}) # examplesIf}
}
\references{
\itemize{
\item https://docs.slack.dev/messaging/sending-and-scheduling-messages
\item https://docs.slack.dev/messaging/sending-and-scheduling-messages#scheduling
}
}
