% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{slack_react}
\alias{slack_react}
\title{Add a reaction to a slack messages}
\usage{
slack_react(ts, emoji, channel)
}
\arguments{
\item{ts}{Numeric. Timestamp to identify the messages}

\item{emoji}{Character vector. Name(s) of emoji's to add as reactions.}

\item{channel}{Character vector. Channel where message can be found.}
}
\value{
Success message
}
\description{
Add a reaction to a slack messages
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
m <- slack_messages("testing-api")
slack_react(m$ts[1], c("bee", "hibiscus", "swan", "ant"), channel = "testing-api")
\dontshow{\}) # examplesIf}
}
