% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{slack_scheduled_list}
\alias{slack_scheduled_list}
\title{List currently scheduled messages}
\usage{
slack_scheduled_list()
}
\value{
Data frame of currently scheduled messages
}
\description{
Returns the currently scheduled messages with added details regarding
timezones, etc.
}
\examples{
slack_scheduled_list()
}
