% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{slack_user}
\alias{slack_user}
\title{Fetch details on a specific users}
\usage{
slack_user(name, users = NULL)
}
\arguments{
\item{name}{Character. String to match real name to}

\item{users}{Data frame. Data frame of users from \code{slack_users()}.}
}
\value{
Data frame
}
\description{
Fetch details on a specific users
}
\examples{
slack_user("Steffi")
}
