% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{slack_users}
\alias{slack_users}
\title{Fetch a list of Slack users}
\usage{
slack_users()
}
\value{
Data frame of Slack users including names and ids
}
\description{
Fetch a list of Slack users
}
\examples{
u <- slack_users()
}
