% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socials.R
\name{socials_post_issue}
\alias{socials_post_issue}
\title{DEPRECATED - Create a draft issue to post to Mastodon and LinkedIn}
\usage{
socials_post_issue(
  time,
  tz = "America/Winnipeg",
  title,
  body,
  where = "mastodon",
  avoid_dups = TRUE,
  add_hash = TRUE,
  dry_run = FALSE,
  open_browser = TRUE,
  over_char_limit = cli::cli_abort,
  verbose = FALSE
)
}
\arguments{
\item{time}{Date/time. Date and time at which the post should be made}

\item{tz}{Character. Timezone (from \code{OlsonNames()}) in which to post}

\item{title}{Character. Title of the post (\verb{[Post]} and the date will be
prepended and appended}

\item{body}{Character. Text to be posted (omit the YAML for posting info;
#RStats and @rstats@a.gup.pe will be appended for Mastodon, #RStats for
LinkedIn),
\emph{or} link to text file with both Mastodon and LinkedIn body text, headed by
by ---- Mastodon ----- and --- LinkedIn -----.}

\item{where}{Character vector. Either \code{mastodon} and/or \code{linkedin} to
specify which platforms this should be posted on.}

\item{avoid_dups}{Logical. Don't post an issue if any open issue has the
same title.}

\item{add_hash}{Logical. Whether to automatically add the RStats hashtags.}

\item{dry_run}{Logical. Whether to perform a dry run (do not post, but
display draft if \code{verbose = TRUE}).}

\item{open_browser}{Logical. Whether to open the issue in the browser.}

\item{over_char_limit}{Function. Error or warn if over the character limit?}

\item{verbose}{Logical. Show extra informative messages.}
}
\description{
Formats the body and title of an issue and posts it on
"rosadmin/scheduled_socials". The issue will bed opened in a browser for
editing and confirmation. Note that issues will not be posted until the
labels "draft" and "needs-review" have been removed.
}
\examples{
# DEPRECATED
# socials_post_issue("2025-01-01 00:00:00",
#                    title = "test post",
#                    body = "Test body",
#                    dry_run = TRUE, verbose = TRUE)
}
