% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/throwback_thursdays.R
\name{tt_post}
\alias{tt_post}
\title{Create Throwback Thursday post}
\usage{
tt_post(
  date,
  title,
  url,
  blurb = "<-- Interesting description/blurb mentioning the author -->",
  dry_run = FALSE,
  print = FALSE
)
}
\arguments{
\item{date}{YMD Character. Date of the original post.}

\item{title}{Character. Title of the original post.}

\item{url}{Character. URL to original blog post}

\item{blurb}{Character. Interesting bit to describe post (optional).}

\item{dry_run}{Logical. Whether to do a dry run (i.e. don't talk to the API).}

\item{print}{Logical. Whether to simply print the text to console instead of
copying to the clipboard.}
}
\value{
Character string of R functions with relevant details (normally
pasted into a POSTS_TT.R script file and tweaked as necessary before being
executed).
}
\description{
Create Throwback Thursday post
}
\examples{
# Standard
tt_post(
  "2019-05-14",
  "POWER to the People",
  "https://ropensci.org/blog/2019/05/14/nasapower/",
  print = TRUE
)

# Double throwback
tt_post(
  c("2017-08-22", "2017-08-22"),
  c("So you (don't) think you can review a package",
    "Onboarding visdat, a tool for preliminary visualisation of whole dataframes"
  ),
  url = c(
    "https://ropensci.org/blog/2017/08/22/first-package-review/",
    "https://ropensci.org/blog/2017/08/22/visdat/"
  ),
  print = TRUE
)

# Slug only
tt_post("2019-05-14", "POWER to the People", "nasapower", print = TRUE)
}
