% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/throwback_thursdays.R
\name{tt_review}
\alias{tt_review}
\title{Review blogposts views for Throwback Thursday highlights}
\usage{
tt_review(which_month = NULL)
}
\arguments{
\item{which_month}{Numeric. Which month to return? Numeric 1-12 for month.
Default \code{NULL} returns all.}
}
\value{
Data frame of posts by views with urls
}
\description{
Use \code{\link[=matomo_update]{matomo_update()}} and friends to download and update blog post view data.
\code{tt_review()} summarizes the posts by view and optionally filters to a
specific month.
}
\examples{
\dontshow{if (dir.exists(cache_dir("matomo"))) withAutoprint(\{ # examplesIf}
tt_review(11)
\dontshow{\}) # examplesIf}
}
