% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecases.R
\name{uc_fetch}
\alias{uc_fetch}
\title{Fetch and format use cases from a GitHub Discussion board}
\usage{
uc_fetch(owner = "ropensci", name = "discussions")
}
\arguments{
\item{owner}{Character. Repository owner.}

\item{name}{Character. Repository for discussions}
}
\value{
Data frame of usecases
}
\description{
Fetch and format use cases from a GitHub Discussion board
}
\examples{
u <- uc_fetch()
}
