% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecases.R
\name{uc_fmt}
\alias{uc_fmt}
\title{Extract out resource information}
\usage{
uc_fmt(uc, min_date, pkgs = NULL)
}
\arguments{
\item{uc}{Data frame. Data frame of usecases from \code{uc_fetch()}.}

\item{min_date}{Character/Date. Minimum date of usecases to retain.}

\item{pkgs}{Data frame. Optional data frame of package details (defaults to \code{pkgs()}).}
}
\value{
Data frame of formatted use cases.
}
\description{
Extract out resource information
}
\examples{
u <- uc_fetch() |>
  uc_fmt("2025-01-01")
}
