% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecases.R
\name{uc_handles}
\alias{uc_handles}
\title{Add social media handles}
\usage{
uc_handles(uc, force_masto = FALSE)
}
\arguments{
\item{uc}{Data frame. Formatted use cases output of \code{uc_fmt()}.}

\item{force_masto}{Logical. Passed to \code{\link[monarch:add_handles]{monarch::add_handles()}}. Whether or
not to force a re-check of mastodon handles (good if you think they've
changed or they are 'none' and you want to check again).}
}
\value{
Data frame with social media handles appended/filled in
}
\description{
Add social media handles
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
u <- uc_fetch() |>
  uc_fmt("2025-01-01") |>
  uc_handles()
\dontshow{\}) # examplesIf}
}
