% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecases.R
\name{uc_names}
\alias{uc_names}
\title{Add in author and maintainer social media handles}
\usage{
uc_names(uc)
}
\arguments{
\item{uc}{Data frame of formatted use cases from \code{uc_fmt()}.}
}
\value{
Data frame of formatted use cases including social media handles.
}
\description{
Add in author and maintainer social media handles
}
\examples{
u <- uc_fetch() |>
  uc_fmt("2025-01-01") |>
  uc_names()
}
