% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecases.R
\name{uc_platform}
\alias{uc_platform}
\title{Arrange by platform}
\usage{
uc_platform(uc)
}
\arguments{
\item{uc}{Data frame. Formatted use cases including social media handles,
output of \code{uc_names()}.}
}
\value{
Data frame of use cases arranged by platform on which to advertise.
}
\description{
Arrange by platform
}
\examples{
u <- uc_fetch() |>
  uc_fmt("2025-01-01") |>
  uc_names() |>
  uc_platform()
}
