% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{url_from_api}
\alias{url_from_api}
\title{Convert an api url to a user-facing url}
\usage{
url_from_api(url)
}
\arguments{
\item{url}{Character. URL to convert}
}
\value{
User-facing url character string
}
\description{
When working with the GitHub api, urls are often in for the api.
}
\examples{
url_from_api("https://api.github.com/repos/ropensci/qualtRics/issues/386")
}
