% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{url_from_path}
\alias{url_from_path}
\title{Create url from content date and slug}
\usage{
url_from_path(
  path,
  date = NULL,
  lang = NULL,
  where = "blog",
  base_url = "https://ropensci.org"
)
}
\arguments{
\item{path}{Character. Slug, URL, or path to file in repository}

\item{date}{Character. Date for event, used to create link (otherwise extracted from slug)}

\item{lang}{Character. Language of blogpost (i.e., \code{en}, \code{es}, \code{fr}, etc.)}

\item{where}{Character. 'blog' or 'event' depending on the content type.}

\item{base_url}{Character. Base url of blog posts.}
}
\value{
Full url
}
\description{
Create url from content date and slug
}
\examples{
url_from_path("my-post", date = "2025-01-01")
url_from_path("2025-01-01-my-post/index.es.md")
url_from_path("content/blog/2025-09-29-news-september-2025/index.md")
url_from_path("content/blog/2025-09-29-news-september-2025/index.Rmd")
url_from_path("content/blog/2025-09-29-news-september-2025/")
}
