% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editorial.R
\name{wordlist_create}
\alias{wordlist_create}
\title{Create Word lists for spelling dictionaries}
\usage{
wordlist_create(
  which = c("packages", "maintainers"),
  path = ".wordlists/names.txt"
)
}
\arguments{
\item{which}{Character vector. Which words to create lists for? Any of
"packages", "maintainers", "contributors".}

\item{path}{Character. File path to save wordlist to.}
}
\value{
Path of new wordlist
}
\description{
Creates a word list of packages, and/or maintainer, and/or contributor names.
If wordlist already exists will update existing list.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
wordlist_create()
\dontshow{\}) # examplesIf}
}
