% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editorial.R
\name{wordlist_update}
\alias{wordlist_update}
\title{Add specific words to the wordlist}
\usage{
wordlist_update(words, path = ".wordlist/names.txt")
}
\arguments{
\item{words}{Character vector. Words to add to word list, note that these are
all split into individual words by spaces.}

\item{path}{Character. Where the word list is to be saved/updated.}
}
\value{
Path of word list
}
\description{
Add specific words to the wordlist
}
