% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yaml_extract}
\alias{yaml_extract}
\title{Extract YAML keys from block}
\usage{
yaml_extract(yaml, trim = "~~~")
}
\arguments{
\item{yaml}{Character. String from which to extract YAML keys}

\item{trim}{Character. Text to remove from the YAML block before processing.
Usually the text that defines the block.}
}
\value{
data frame of yaml keys
}
\description{
Extract YAML keys from block
}
\examples{

yaml_extract("~~~start: 2023-11-12\nauthor: Steffi\n~~~")

}
