% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{buffer_channels}
\alias{buffer_channels}
\title{Return Buffer channel ids}
\usage{
buffer_channels(org = buff_org, dry_run = FALSE)
}
\arguments{
\item{org}{Character. Organizational ID to return channels for (see
\code{\link[=buffer_org]{buffer_org()}}). Uses rOpenSci's by default \code{buff_org}.}
}
\value{
Data frame with \code{id} and \code{service} (linkedin, mastodon, bluesky,
etc.)
}
\description{
Return Buffer channel ids
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
buffer_channels()
\dontshow{\}) # examplesIf}
}
