% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{buffer_posts_list}
\alias{buffer_posts_list}
\title{Return a list of Buffer posts}
\usage{
buffer_posts_list(
  status = c("scheduled", "draft"),
  fields = c("id", "channelService", "text", "createdAt", "dueAt", "status"),
  dry_run = FALSE
)
}
\arguments{
\item{status}{Character vector. Status of posts to return. Defaults to
"scheduled" and "draft", set to "sent" to return all posted messages.
Options are "draft", "error", "needs_approal","scheduled", "sending",
"sent".}

\item{filds}{Character vector. Information to return. Become the columns in
the data frame.}
}
\description{
Fetches a list of posts on all channels on Buffer.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
buffer_posts_list()
buffer_posts_list(status = "error")
buffer_posts_list(status = "sent")
\dontshow{\}) # examplesIf}
}
\references{
Post status: https://developers.buffer.com/reference.html#type/PostStatus
Posts: https://developers.buffer.com/reference.html#field-posts
Fields: https://developers.buffer.com/reference.html#type/Post
}
