% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.R
\name{buffer_posts_write}
\alias{buffer_posts_write}
\title{Create a social media post on Buffer}
\usage{
buffer_posts_write(
  body,
  when,
  tz = NULL,
  channels = c("mastodon", "linkedin", "bluesky"),
  draft = TRUE,
  dry_run = FALSE,
  open_browser = interactive(),
  handles = NULL,
  ...
)
}
\arguments{
\item{body}{Character vector. Text to use as the body of the post. Can be a
vector the same length as \code{channels} to use different text for each channel.}

\item{when}{Character/Date time. Either "now" to post immediately or create a
draft post for immediate posting, or a Date/time to specify when the post
should be scheduled.}

\item{tz}{Character. The \code{\link[=OlsonNames]{OlsonNames()}} timezone to use for \code{when} if it is a
character string (if \code{when} is a date/time \code{tz} is ignored). Default
(\code{NULL}) uses system timezone.}

\item{channels}{Character vector. Where to post? Options are "mastodon",
"bluesky", and/or "linkedin".}

\item{draft}{Logical. Whether this post should be a draft.}

\item{dry_run}{Logical. Whether to do a dry run (i.e. don't talk to the API).}

\item{open_browser}{Logical. Whether to open the issue in the browser.}

\item{...}{Handles to add to the body. Should be \code{id = c(mastodon = "handle", bluesky = "handle", linkedin = "handle")} where id is used in the body as
\code{{id}} and the names (mastodon, bluesky, linkedin) must match the channels.}
}
\value{
Data frame of post ids
}
\description{
Create a social media post on Buffer
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
p <- buffer_posts_write("testing Api again...", when = "now") # Create draft
buffer_posts_remove(p$id)
p <- buffer_posts_write("testing Api again...", when = "2027-01-01 10:00") # Create schedueld draft
buffer_posts_remove(p$id)

# Create draft using appropriate handles
p <- buffer_posts_write(
  "testing Api again... {steffi} and {yani}",
  when = "now",
  steffi = c(
    "mastodon" = "@steffilazerte@fosstodon.org",
    "bluesky" = "@steffilazerte.bsky.social",
    "linkedin" = "Steffi LaZerte"
  ),
  yani = c(
    "mastodon" = "@yabellini@rstats.me",
    "bluesky" = "@yabellini.bsky.social",
    "linkedin" = "Yanina Bellini Saibene"
  )
)

buffer_posts_remove(p$id)

p <- buffer_posts_write(
  paste0(rep("testing Api again...\n\n", 20), collapse = ""),
  when = "now"
)
\dontshow{\}) # examplesIf}
}
