% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_utils.R
\name{buffer_query}
\alias{buffer_query}
\title{Great GraphQL Query for Buffer}
\usage{
buffer_query(template, fields = NULL, sort = NULL, filter = NULL, ...)
}
\arguments{
\item{fields}{Character vector. Fields to return}

\item{...}{Character or lists. Query Parameters}

\item{type}{Character. Query recipe 'org', 'schedule', 'remove', 'write'.}
}
\description{
Note that the GraphQL requires " (' won't work).
}
\examples{
buffer_query(type = "org", fields = c("id", "name"))

buffer_query(
  type = "scheduled",
  fields = c("id", "text", "createdAt"),
  sort = list("dueAt" = "asc", "createdAt" = "desc"),
  filter = list("status" = "scheduled"),
  org = ro_org
)
}
