% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_dir}
\alias{cache_dir}
\title{Cache folder to store data}
\usage{
cache_dir(type)
}
\value{
Character file path
}
\description{
Cache folder to store data
}
\examples{
cache_dir(type = "matomo")
cache_dir(type = "help-wanted")
}
