% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_common_docs.R
\name{common_docs}
\alias{common_docs}
\title{Common arguments and documentation for various functions}
\arguments{
\item{test_run}{Logical. Whether to do a test run (i.e. post to a test area)}

\item{dry_run}{Logical. Whether to do a dry run (i.e. don't post)}

\item{quiet}{Logical. Whether to suppress progress messages.}

\item{verbose}{Logical. Show extra informative messages.}
}
\description{
Common arguments and documentation for various functions
}
\details{
Use \verb{@inheritParams common_docs} to include the above in any function
documentation with a matching argument (will only include matching args)
}
\keyword{internal}
