% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coworking.R
\name{cw_event}
\alias{cw_event}
\title{Create a draft event for coworking}
\usage{
cw_event(date, dry_run = FALSE)
}
\arguments{
\item{date}{Character/Date. Date of the coworking event (local)}

\item{dry_run}{Logical. Whether to really create the event or just return the
text.}
}
\description{
Creates a draft coworking event for the roweb3 website. All details pulled
from the coworking todo list issue in \code{rosadmin/comms}.
}
