% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coworking.R
\name{cw_issue}
\alias{cw_issue}
\title{Create Coworking To-Do's on Comms Repo}
\usage{
cw_issue(
  date = NULL,
  tz = NULL,
  theme = "XXXX",
  cohost = "XXXX",
  dry_run = FALSE
)
}
\arguments{
\item{date}{Character. Date of next event (if NULL picks next first Tuesday).}

\item{tz}{Character. Timezone of next event (if NULL picks next in order).}

\item{theme}{Character. Name of theme, if unknown, uses XXXX placeholder}

\item{cohost}{Character. Name of cohost, if unknown, uses XXXX placeholder}
}
\description{
Create an issue listing the coworking todos. If no date or timezone, picks
the next appropriate date (first Tuesday in the month following an existing
coworking issue) and next appropriate timezone (cycling through America,
Europe, and Australia) automatically.
}
