% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coworking.R
\name{cw_socials}
\alias{cw_socials}
\title{Create a draft post for coworking}
\usage{
cw_socials(
  date,
  who_masto,
  who_slack,
  who_linkedin,
  posters_tz = "America/Winnipeg",
  dry_run = FALSE
)
}
\arguments{
\item{date}{Character/Date. Date of the coworking event (local)}

\item{who_masto}{Character. The full mastodon handle for the cohost (i.e. XXXX@XXXX.com)}

\item{who_slack}{Character. The full Slack handle for the cohost (i.e. @XXXX)}

\item{who_linkedin}{Character. The full LinkedIn handle for the cohost (i.e. @XXXX)}

\item{posters_tz}{Character. Timezone of poster. Required for getting the
time at which to post Slack messages as these are posted in the local
timezone}

\item{when}{Character. Is this a "week" before reminder or an "hour" before reminder?}
}
\description{
Creates draft posts for Mastodon and LinkedIn (by opening issues on
rosadmin/scheduled_socials) and Slack (by printing the post text and
schedule).
}
\examples{

\dontrun{
cw_socials("2023-07-04", who_masto = "@cohost@mastodon.org", who_slack = "@cohost")
}
}
