% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coworking.R
\name{cw_socials}
\alias{cw_socials}
\title{Create a draft post for coworking}
\usage{
cw_socials(
  date,
  who_masto,
  who_slack,
  who_linkedin,
  who_main_masto = "@steffilazerte@fosstodon.org",
  who_main_slack = "<@UNRAUCMTK>",
  who_main_linkedin = "Steffi LaZerte",
  posters_tz = "America/Winnipeg",
  dry_run = FALSE,
  branch = NULL
)
}
\arguments{
\item{date}{Character/Date. Date of the coworking event (local)}

\item{who_masto}{Character. The full mastodon handle for the cohost (i.e. XXXX@XXXX.com)}

\item{who_slack}{Character. The full Slack handle for the cohost (i.e. @XXXX)}

\item{who_linkedin}{Character. The full LinkedIn handle for the cohost (i.e. @XXXX)}

\item{who_main_masto}{Character. The full mastodon handle for the rOpenSci staff organizer.}

\item{who_main_slack}{Character. The Slack id for the rOpenSci staff
organizer (i.e., \verb{<@UXXXXXXX>}. Defaults to Steffi's id.}

\item{who_main_linkedin}{Character. The full LinkedIn handle for the rOpenSci staff organizer.}

\item{posters_tz}{Character. Timezone of poster. Required for getting the
time at which to post Slack messages as these are posted in the local
timezone}

\item{dry_run}{Logical. Whether to do a dry run (i.e. don't post)}

\item{branch}{Character. Branch name if not on main.}
}
\description{
Creates draft posts for Mastodon and LinkedIn (by opening issues on
rosadmin/scheduled_socials) and Slack (by printing the post text and
schedule).
}
\examples{

\dontrun{
cw_socials("2023-07-04", who_masto = "@cohost@mastodon.org", who_slack = "<UXXXXX>")
}
}
