% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionaries.R
\name{dict_usecases}
\alias{dict_usecases}
\title{Dictionary for usecases}
\usage{
dict_usecases()
}
\value{
Data frame.
}
\description{
Creates a data frame dictionary of different versions of wording to use in
usecase posts depending on the language of the post.
}
\examples{
dict_usecases()
}
