% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_wanted.R
\name{help_fetch}
\alias{help_fetch}
\title{Fetch help wanted issues}
\usage{
help_fetch(min_date, json = "/repos/rosadmin/help-wanted/contents/issues.json")
}
\value{
Data frame of help wanted issues
}
\description{
Fetch help wanted issues
}
\examples{
h <- help_fetch("2025-01-01")
}
