% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_wanted.R
\name{help_handles}
\alias{help_handles}
\title{Title}
\usage{
help_handles(help, force_masto = FALSE)
}
\arguments{
\item{help}{}
}
\description{
Title
}
\examples{
h <- help_fetch("2025-01-01") |>
  help_handles()
}
