% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_wanted.R
\name{help_post}
\alias{help_post}
\title{Create help-wanted post}
\usage{
help_post(help, skip = NULL, date_time = NULL, dry_run = FALSE)
}
\arguments{
\item{help}{Data frame. Formatted help-wanted issues including social media
handles.}

\item{skip}{Numeric vector. The help wanted issues to skip creating a post
for.}

\item{date_time}{Character/Date. When to post. Defaults to next Thursday if
\code{NULL}.}

\item{dry_run}{Logical. Whether to do a dry run (i.e. don't talk to the API).}
}
\value{
Copies commands to clipboard, optionally prints if (\code{print = TRUE}).
}
\description{
Create help-wanted post
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
h <- help_fetch()

help_preview(h)
# Review and note issues to skip

help_post(h, skip = c(1:6), dry_run = TRUE)  # For non-interactive examples
\dontshow{\}) # examplesIf}
}
