% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_wanted.R
\name{help_post}
\alias{help_post}
\title{Create help-wanted \code{socials_post_issue()} command}
\usage{
help_post(help, date_time = NULL, dry_run = FALSE, print = FALSE)
}
\arguments{
\item{help}{Data frame. Formatted help-wanted issues including social media
handles and arranged by platform upon which to post, output of
\code{by_platform()}.}

\item{date_time}{Character/Date. When to post. Defaults to next Thursday if \code{NULL}.}

\item{dry_run}{Logical. Whether to do a dry run (i.e. don't talk to the API).}

\item{print}{Logical. Whether to simply print the text to console instead of
copying to the clipboard.}
}
\value{
Copies commands to clipboard, optionally prints if (\code{print = TRUE}).
}
\description{
Create help-wanted \code{socials_post_issue()} command
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
h <- help_fetch("2025-05-23") |>
  help_handles() |>
  by_platform()

help_post(h, print = TRUE)  # For non-interactive examples
\dontshow{\}) # examplesIf}
}
