% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_wanted_json.R
\name{hw_issues}
\alias{hw_issues}
\title{Create help wanted JSON file}
\usage{
hw_issues(verbose = TRUE)
}
\value{
Creates/updates \code{issues.json}
}
\description{
Function created for rosadmin/help-wanted. Fetches package repos with
issues labelled help-wanted. Saves output to issues.json for use by
helpwanted workflows including https://ropensci.org/help-wanted.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
hw_issues()
\dontshow{\}) # examplesIf}
}
