% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedin.R
\name{li_posts_read}
\alias{li_posts_read}
\title{Get a list of recent posts by rOpenSci}
\usage{
li_posts_read(author)
}
\arguments{
\item{author}{Character. LinkedIn URN id (e.g., rOpenSci's is "urn:li:organization:77132573")}
}
\value{
list of posts
}
\description{
Get a list of recent posts by rOpenSci
}
\examples{

li_posts_read(ro_urn)
}
\references{
\itemize{
\item https://learn.microsoft.com/en-us/linkedin/shared/api-guide/concepts/urns
\item https://learn.microsoft.com/en-us/linkedin/marketing/integrations/community-management/shares/posts-api
}
}
