% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkg_authors}
\alias{pkg_authors}
\title{Get package author names}
\usage{
pkg_authors(package, pkgs)
}
\arguments{
\item{package}{Character. Package name}

\item{pkgs}{Data frame. Packages returned by \code{\link[=pkgs_ru]{pkgs_ru()}}.}
}
\value{
Character name of maintainer
}
\description{
Get package author names
}
\examples{
pkg_authors("weathercan", pkgs_ru())
}
