% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkg_authors}
\alias{pkg_authors}
\title{Get package author names}
\usage{
pkg_authors(x, pkgs)
}
\arguments{
\item{x}{Character. Package name}

\item{pkgs}{Data frame. Packages returned by \code{pkgs()}.}
}
\value{
Character name of maintainer
}
\description{
Get package author names
}
