% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promoutils-package.R
\docType{package}
\name{promoutils-package}
\alias{promoutils}
\alias{promoutils-package}
\title{promoutils: Utilities for Promoting rOpenSci on Twitter}
\description{
Utility functions for accessing GitHub and twitter data.
}
\author{
\strong{Maintainer}: Steffi LaZerte \email{steffi@steffi.ca} (\href{https://orcid.org/0000-0002-7690-8360}{ORCID})

}
\keyword{internal}
