% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promoutils-package.R
\docType{package}
\name{promoutils-package}
\alias{promoutils}
\alias{promoutils-package}
\title{promoutils: Utilities for Promoting rOpenSci on Social Media}
\description{
Utility functions for accessing GitHub and social media data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci-org/promoutils}
  \item Report bugs at \url{https://github.com/ropensci-org/promoutils/issues}
}

}
\author{
\strong{Maintainer}: Steffi LaZerte \email{steffi@steffi.ca} (\href{https://orcid.org/0000-0002-7690-8360}{ORCID})

}
\keyword{internal}
