% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{slack_channels}
\alias{slack_channels}
\title{Title}
\usage{
slack_channels(types = c("public_channel", "private_channel"))
}
\arguments{
\item{channel}{}
}
\description{
Title
}
\examples{
slack_channels()
slack_channels("general")
chn <- slack_channels(types = "private_channel")
}
