% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{slack_message_rm}
\alias{slack_message_rm}
\title{Remove a Slack message}
\usage{
slack_message_rm(msg = NULL, channel = NULL, channel_id = NULL, ts = NULL)
}
\arguments{
\item{msg}{Data frame. Output of \code{slack_messages} containing message to
remove. Should contain columns "channel" and "ts". Note that only the most
recent message will be removed.}

\item{channel}{Character. Channel Name. Not required if channel_id supplied.}

\item{channel_id}{Character. Channel id.}

\item{ts}{Numeric. Timestamp to identify message to remove (if no \code{msg})}
}
\value{
Nothing. Side effect of removing message from channel
}
\description{
Use with caution! Note that users can only delete their own messages. If you
try to delete another user's message you will get a "Cant Delete Message"
message from the Slack API.
}
\seealso{
\code{\link[=slack_message_rm_bulk]{slack_message_rm_bulk()}}
}
