% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{slack_messages}
\alias{slack_messages}
\title{Get the last 100 messages from a channel}
\usage{
slack_messages(channel = NULL, channel_id = NULL)
}
\arguments{
\item{channel}{Character. Channel Name. Not required if channel_id supplied.}

\item{channel_id}{Character. Channel id.}
}
\value{
Data frame. Messages and details
}
\description{
Requires one of \code{channel} OR \code{channel_id}
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
slack_messages(channel_id = "C026GCWKA") # General
slack_messages(channel = "General")
\dontshow{\}) # examplesIf}
}
