% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack.R
\name{slack_messages}
\alias{slack_messages}
\title{Get the last 100 messages from a channel}
\usage{
slack_messages(channel = NULL, channel_id = NULL)
}
\description{
Get the last 100 messages from a channel
}
\examples{
slack_messages(channel_id = "C026GCWKA") # General
}
