% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecases.R
\name{uc_handles}
\alias{uc_handles}
\title{Add social media handles}
\usage{
uc_handles(uc)
}
\arguments{
\item{uc}{Data frame. Formatted use cases output of \code{uc_fmt()}.}
}
\value{
Data frame with social media handles appended/filled in
}
\description{
Add social media handles
}
\examples{
u <- uc_fetch() |>
  uc_fmt("2025-01-01") |>
  uc_handles()
}
