% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecases.R
\name{uc_post}
\alias{uc_post}
\title{Create use case \code{socials_post_issue()} command}
\usage{
uc_post(uc, date_time = NULL, dry_run = FALSE, print = FALSE)
}
\arguments{
\item{uc}{Data frame. Formatted use cases including social media handles
and arranged by platform upon which to post, output of \code{uc_platform()}.}

\item{date_time}{Character/Date. When to post. Defaults to next Wednesday if \code{NULL}.}

\item{dry_run}{Logical. Whether to do a dry run (i.e. don't post)}

\item{print}{Logical. Whether to simply print the text to console instead of
copying to the clipboard.}
}
\value{
Copies commands to clipboard, optionally prints if (\code{print = TRUE}).
}
\description{
Create use case \code{socials_post_issue()} command
}
\examples{
u <- uc_fetch() |>
  uc_fmt("2025-01-01") |>
  uc_handles() |>
  by_platform()

uc_post(u, dry_run = TRUE)
}
